test(ui-dialog,ui-position): migrate old Position tests
Closes: INSTUI-4051
