--- packages/ui-position/src/Position/__tests__/Position.test.tsx
@@ -1,707 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2015 - present Instructure, Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-import React from 'react'
-import { expect, mount, wait, stub } from '@instructure/ui-test-utils'
-import { within } from '@instructure/ui-utils'
-
-import { Position } from '../index'
-import { PositionLocator } from '../PositionLocator'
-import { PlacementPropValues } from '../../PositionPropTypes'
-
-describe('<Position />', async () => {
-  const parentDefaults = {
-    width: 500,
-    height: 150,
-    padding: 100,
-    overflow: 'auto'
-  }
-
-  it('should render', async () => {
-    await mount(
-      <div style={{ padding: '50px' }}>
-        <div style={{ ...parentDefaults }}>
-          <Position constrain="window" renderTarget={<button>Target</button>}>
-            <div id="content">
-              <div>Content</div>
-            </div>
-          </Position>
-        </div>
-      </div>
-    )
-
-    const position = await PositionLocator.find()
-    const target = await position.findTarget(':contains(Target)')
-    const content = await position.findContent(':contains(Content)')
-
-    expect(target).to.exist()
-    expect(content).to.exist()
-  })
-
-  it('should absolutely position content', async () => {
-    await mount(
-      <div style={{ padding: '50px' }}>
-        <div style={{ ...parentDefaults }}>
-          <Position constrain="window" renderTarget={<button>Target</button>}>
-            <div id="content">
-              <div>Content</div>
-            </div>
-          </Position>
-        </div>
-      </div>
-    )
-
-    const position = await PositionLocator.find()
-    const content = await position.findContent(':contains(Content)')
-    expect((content.getDOMNode() as HTMLDivElement).style.position).to.equal(
-      'absolute'
-    )
-  })
-
-  it('should render right of target', async () => {
-    const onPositionChanged = stub()
-    await mount(
-      <div style={{ padding: '50px' }}>
-        <div style={{ ...parentDefaults }}>
-          <Position
-            placement="end"
-            onPositionChanged={onPositionChanged}
-            renderTarget={<button>Target</button>}
-          >
-            <div id="content">
-              <div>Content</div>
-            </div>
-          </Position>
-        </div>
-      </div>
-    )
-
-    const position = await PositionLocator.find()
-    const target = await position.findTarget(':contains(Target)')
-    const content = await position.findContent(':contains(Content)')
-    // wait for positioning
-    await wait(() => {
-      expect(onPositionChanged).to.have.been.called()
-    })
-
-    const targetRect = target.getBoundingClientRect()
-    const contentRect = content.getBoundingClientRect()
-
-    expect(
-      within(Math.floor(contentRect.left), Math.floor(targetRect.right))
-    ).to.be.true()
-  })
-
-  it('should render top stretched inside of target', async () => {
-    const onPositionChanged = stub()
-    await mount(
-      <div style={{ padding: '50px' }}>
-        <div style={{ ...parentDefaults }}>
-          <Position
-            placement="top stretch"
-            shouldPositionOverTarget
-            onPositionChanged={onPositionChanged}
-            renderTarget={<button>Target</button>}
-          >
-            <div id="content">
-              <div>Content</div>
-            </div>
-          </Position>
-        </div>
-      </div>
-    )
-
-    const position = await PositionLocator.find()
-    const target = await position.findTarget(':contains(Target)')
-    const content = await position.findContent(':contains(Content)')
-    // wait for positioning
-    await wait(() => {
-      expect(onPositionChanged).to.have.been.called()
-    })
-
-    const targetRect = target.getBoundingClientRect()
-    const contentRect = content.getBoundingClientRect()
-
-    expect(
-      within(Math.floor(contentRect.top), Math.floor(targetRect.top), 1)
-    ).to.be.true()
-    expect(
-      within(Math.floor(contentRect.width), Math.floor(targetRect.width), 1)
-    ).to.be.true()
-  })
-
-  it('should render below target', async () => {
-    const onPositionChanged = stub()
-    await mount(
-      <div style={{ padding: '50px' }}>
-        <div style={{ ...parentDefaults }}>
-          <Position
-            placement="bottom"
-            onPositionChanged={onPositionChanged}
-            renderTarget={<button>Target</button>}
-          >
-            <div id="content">
-              <div>Content</div>
-            </div>
-          </Position>
-        </div>
-      </div>
-    )
-
-    const position = await PositionLocator.find()
-    const target = await position.findTarget(':contains(Target)')
-    const content = await position.findContent(':contains(Content)')
-    // wait for positioning
-    await wait(() => {
-      expect(onPositionChanged).to.have.been.called()
-    })
-
-    const targetRect = target.getBoundingClientRect()
-    const contentRect = content.getBoundingClientRect()
-
-    expect(
-      within(Math.floor(contentRect.top), Math.floor(targetRect.bottom), 1)
-    ).to.be.true()
-  })
-
-  it('should render left of target', async () => {
-    const onPositionChanged = stub()
-    await mount(
-      <div style={{ padding: '50px' }}>
-        <div style={{ ...parentDefaults }}>
-          <Position
-            placement="start"
-            onPositionChanged={onPositionChanged}
-            renderTarget={<button>Target</button>}
-          >
-            <div id="content">
-              <div>Content</div>
-            </div>
-          </Position>
-        </div>
-      </div>
-    )
-
-    const position = await PositionLocator.find()
-    const target = await position.findTarget(':contains(Target)')
-    const content = await position.findContent(':contains(Content)')
-    // wait for positioning
-    await wait(() => {
-      expect(onPositionChanged).to.have.been.called()
-    })
-
-    const targetRect = target.getBoundingClientRect()
-    const contentRect = content.getBoundingClientRect()
-
-    expect(
-      within(Math.floor(contentRect.right), Math.floor(targetRect.left))
-    ).to.be.true()
-  })
-
-  it('should render above target', async () => {
-    const onPositionChanged = stub()
-    await mount(
-      <div style={{ padding: '50px' }}>
-        <div style={{ ...parentDefaults }}>
-          <Position
-            placement="top"
-            onPositionChanged={onPositionChanged}
-            renderTarget={<button>Target</button>}
-          >
-            <div id="content">
-              <div>Content</div>
-            </div>
-          </Position>
-        </div>
-      </div>
-    )
-
-    const position = await PositionLocator.find()
-    const target = await position.findTarget(':contains(Target)')
-    const content = await position.findContent(':contains(Content)')
-    // wait for positioning
-    await wait(() => {
-      expect(onPositionChanged).to.have.been.called()
-    })
-
-    const targetRect = target.getBoundingClientRect()
-    const contentRect = content.getBoundingClientRect()
-
-    expect(Math.floor(contentRect.bottom)).to.equal(Math.floor(targetRect.top))
-  })
-
-  it('should center vertically', async () => {
-    const onPositionChanged = stub()
-    await mount(
-      <div style={{ padding: '50px' }}>
-        <div style={{ ...parentDefaults }}>
-          <Position
-            placement="end"
-            onPositionChanged={onPositionChanged}
-            renderTarget={<button>Target</button>}
-          >
-            <div id="content">
-              <div>Content</div>
-            </div>
-          </Position>
-        </div>
-      </div>
-    )
-
-    const position = await PositionLocator.find()
-    const target = await position.findTarget(':contains(Target)')
-    const content = await position.findContent(':contains(Content)')
-    // wait for positioning
-    await wait(() => {
-      expect(onPositionChanged).to.have.been.called()
-    })
-
-    const targetRect = target.getBoundingClientRect()
-    const contentRect = content.getBoundingClientRect()
-
-    const top = Math.floor(contentRect.top)
-    const center = Math.floor(
-      targetRect.top + (targetRect.height / 2 - contentRect.height / 2)
-    )
-
-    expect(within(top, center)).to.be.true()
-  })
-
-  it('should center horizontally', async () => {
-    const onPositionChanged = stub()
-    await mount(
-      <div style={{ padding: '50px' }}>
-        <div style={{ ...parentDefaults }}>
-          <Position
-            placement="bottom"
-            onPositionChanged={onPositionChanged}
-            renderTarget={<button>Target</button>}
-          >
-            <div id="content">
-              <div>Content</div>
-            </div>
-          </Position>
-        </div>
-      </div>
-    )
-
-    const position = await PositionLocator.find()
-    const target = await position.findTarget(':contains(Target)')
-    const content = await position.findContent(':contains(Content)')
-    // wait for positioning
-    await wait(() => {
-      expect(onPositionChanged).to.have.been.called()
-    })
-
-    const targetRect = target.getBoundingClientRect()
-    const contentRect = content.getBoundingClientRect()
-
-    const left = Math.floor(contentRect.left)
-    const targetCenter = targetRect.width / 2
-    const contentCenter = contentRect.width / 2
-    const center = Math.floor(targetRect.left + (targetCenter - contentCenter))
-
-    expect(within(left, center)).to.be.true()
-  })
-
-  describe('with offset props', () => {
-    function assertOffset(
-      placement: PlacementPropValues,
-      offset: number | string,
-      assertion: (newContentRect: DOMRect, top: number, left: number) => void
-    ) {
-      it(`should render offset for ${placement}`, async () => {
-        const onPositionChanged = stub()
-
-        const subject = await mount(
-          <Position
-            constrain="none"
-            placement={placement}
-            onPositionChanged={onPositionChanged}
-            renderTarget={<button>Target</button>}
-          >
-            <div id="content">
-              <div>Content</div>
-            </div>
-          </Position>
-        )
-
-        const position = await PositionLocator.find()
-        const content = await position.findContent(':contains(Content)')
-
-        await wait(() => {
-          expect(onPositionChanged).to.have.been.called()
-        })
-
-        const contentRect = content.getBoundingClientRect()
-        const { top, left } = contentRect
-        const offsetX = offset
-        const offsetY = offset
-
-        await subject.setProps({ offsetX, offsetY })
-        await wait(() => {
-          expect(onPositionChanged).to.have.been.called()
-        })
-
-        const newContentRect = content.getBoundingClientRect()
-        assertion(newContentRect, top, left)
-      })
-    }
-
-    assertOffset('top', 10, (contentRect, top, left) => {
-      expect(within(contentRect.top, top - 10, 1)).to.be.true()
-      expect(within(contentRect.left, left - 10, 1)).to.be.true()
-    })
-
-    assertOffset('start', '10px', (contentRect, top, left) => {
-      expect(within(contentRect.top, top - 10, 1)).to.be.true()
-      expect(within(contentRect.left, left - 10, 1)).to.be.true()
-    })
-
-    assertOffset('end', 10, (contentRect, top, left) => {
-      expect(within(contentRect.top, top - 10, 1)).to.be.true()
-      expect(within(contentRect.left, left + 10, 1)).to.be.true()
-    })
-
-    assertOffset('bottom', 10, (contentRect, top, left) => {
-      expect(within(contentRect.top, top + 10, 1)).to.be.true()
-      expect(within(contentRect.left, left - 10, 1)).to.be.true()
-    })
-  })
-
-  describe('when constrained to scroll-parent', async () => {
-    it('should re-position below target', async () => {
-      const onPositionChanged = stub()
-
-      await mount(
-        <div style={{ padding: '50px' }}>
-          <div id="mountNode">mount</div>
-          <div
-            style={{
-              width: '50px',
-              height: '50px',
-              overflow: 'scroll',
-              padding: '0 50px 50px 50px'
-            }}
-          >
-            <Position
-              placement="top"
-              constrain="scroll-parent"
-              mountNode={() => document.getElementById('mountNode')}
-              onPositionChanged={onPositionChanged}
-              renderTarget={<button>Target</button>}
-            >
-              <div id="content">
-                <div>Content</div>
-              </div>
-            </Position>
-          </div>
-        </div>
-      )
-
-      const position = await PositionLocator.find()
-      const target = await position.findTarget(':contains(Target)')
-      const content = await position.findContent(':contains(Content)')
-      // wait for positioning
-      await wait(() => {
-        expect(onPositionChanged).to.have.been.called()
-      })
-
-      const targetRect = target.getBoundingClientRect()
-      const contentRect = content.getBoundingClientRect()
-
-      expect(
-        within(Math.floor(contentRect.top), Math.floor(targetRect.bottom), 1)
-      ).to.be.true()
-    })
-
-    it('should re-position above target', async () => {
-      const onPositionChanged = stub()
-
-      await mount(
-        <div style={{ padding: '50px' }}>
-          <div id="mountNode">mount</div>
-          <div
-            style={{
-              width: '50px',
-              height: '0px',
-              overflow: 'scroll',
-              padding: '50px 50px 0 50px'
-            }}
-          >
-            <Position
-              placement="bottom"
-              constrain="scroll-parent"
-              mountNode={() => document.getElementById('mountNode')}
-              onPositionChanged={onPositionChanged}
-              renderTarget={<button>Target</button>}
-            >
-              <div id="content">
-                <div>Content</div>
-              </div>
-            </Position>
-          </div>
-        </div>
-      )
-
-      const position = await PositionLocator.find()
-      const target = await position.findTarget(':contains(Target)')
-      const content = await position.findContent(':contains(Content)')
-      // wait for positioning
-      await wait(() => {
-        expect(onPositionChanged).to.have.been.called()
-      })
-
-      const targetRect = target.getBoundingClientRect()
-      const contentRect = content.getBoundingClientRect()
-
-      expect(
-        within(Math.floor(contentRect.bottom), Math.floor(targetRect.top), 1)
-      ).to.be.true()
-    })
-
-    it('should re-position after target', async () => {
-      const onPositionChanged = stub()
-
-      await mount(
-        <div style={{ padding: '50px' }}>
-          <div id="mountNode">mount</div>
-          <div
-            style={{
-              width: '50px',
-              height: '50px',
-              overflow: 'scroll',
-              padding: '50px 80px 50px 0'
-            }}
-          >
-            <Position
-              placement="start"
-              constrain="scroll-parent"
-              mountNode={() => document.getElementById('mountNode')}
-              onPositionChanged={onPositionChanged}
-              renderTarget={<button>Target</button>}
-            >
-              <div id="content">
-                <div>Content</div>
-              </div>
-            </Position>
-          </div>
-        </div>
-      )
-
-      const position = await PositionLocator.find()
-      const target = await position.findTarget(':contains(Target)')
-      const content = await position.findContent(':contains(Content)')
-      // wait for positioning
-      await wait(() => {
-        expect(onPositionChanged).to.have.been.called()
-      })
-
-      const targetRect = target.getBoundingClientRect()
-      const contentRect = content.getBoundingClientRect()
-
-      expect(
-        within(Math.floor(contentRect.left), Math.floor(targetRect.right), 1)
-      ).to.be.true()
-    })
-
-    it('should re-position before target', async () => {
-      const onPositionChanged = stub()
-
-      await mount(
-        <div style={{ padding: '50px' }}>
-          <div id="mountNode">mount</div>
-          <div
-            style={{
-              width: '50px',
-              height: '50px',
-              overflow: 'scroll',
-              padding: '50px 0px 50px 80px'
-            }}
-          >
-            <Position
-              placement="end"
-              constrain="scroll-parent"
-              mountNode={() => document.getElementById('mountNode')}
-              onPositionChanged={onPositionChanged}
-              renderTarget={<button>Target</button>}
-            >
-              <div id="content">
-                <div>Content</div>
-              </div>
-            </Position>
-          </div>
-        </div>
-      )
-
-      const position = await PositionLocator.find()
-      const target = await position.findTarget(':contains(Target)')
-      const content = await position.findContent(':contains(Content)')
-      // wait for positioning
-      await wait(() => {
-        expect(onPositionChanged).to.have.been.called()
-      })
-
-      const targetRect = target.getBoundingClientRect()
-      const contentRect = content.getBoundingClientRect()
-
-      expect(
-        within(Math.floor(contentRect.right), Math.floor(targetRect.left), 1)
-      ).to.be.true()
-    })
-  })
-
-  describe('when the documentElement is offset', async () => {
-    beforeEach(async () => {
-      document.documentElement.style.position = 'fixed'
-      document.documentElement.style.top = '-100px'
-    })
-
-    it('should position correctly', async () => {
-      await mount(
-        <div style={{ padding: '100px' }}>
-          <Position placement="bottom" renderTarget={<button>Target</button>}>
-            <div id="content">
-              <div>Content</div>
-            </div>
-          </Position>
-        </div>
-      )
-
-      const position = await PositionLocator.find()
-      const target = await position.findTarget(':contains(Target)')
-      const content = await position.findContent(':contains(Content)')
-
-      await wait(() => {
-        expect(content.getBoundingClientRect().top).to.equal(
-          target.getBoundingClientRect().bottom
-        )
-      })
-    })
-
-    it('should position correctly with mountNode', async () => {
-      await mount(
-        <div style={{ padding: '100px' }}>
-          <div id="mountNode">mount</div>
-          <div
-            style={{
-              width: '50px',
-              height: '50px',
-              overflow: 'scroll',
-              padding: '50px'
-            }}
-          >
-            <Position
-              placement="bottom"
-              constrain="scroll-parent"
-              mountNode={() => document.getElementById('mountNode')}
-              renderTarget={<button>Target</button>}
-            >
-              <div id="content">
-                <div>Content</div>
-              </div>
-            </Position>
-          </div>
-        </div>
-      )
-
-      const position = await PositionLocator.find()
-      const target = await position.findTarget(':contains(Target)')
-      const content = await position.findContent(':contains(Content)')
-
-      await wait(() => {
-        expect(content.getBoundingClientRect().top).to.equal(
-          target.getBoundingClientRect().bottom
-        )
-      })
-    })
-  })
-
-  describe('containerDisplay prop', async () => {
-    it('should be <span> default when unset', async () => {
-      await mount(
-        <div style={{ padding: '50px' }}>
-          <div style={{ ...parentDefaults }}>
-            <Position constrain="window" renderTarget={<button>Target</button>}>
-              <div id="content">
-                <div>Content</div>
-              </div>
-            </Position>
-          </div>
-        </div>
-      )
-
-      const position = await PositionLocator.find()
-      const style = getComputedStyle(position.getDOMNode())
-
-      expect(style.display).to.equal('inline')
-    })
-
-    it('should apply "inline-block"', async () => {
-      await mount(
-        <div style={{ padding: '50px' }}>
-          <div style={{ ...parentDefaults }}>
-            <Position
-              constrain="window"
-              renderTarget={<button>Target</button>}
-              containerDisplay="inline-block"
-            >
-              <div id="content">
-                <div>Content</div>
-              </div>
-            </Position>
-          </div>
-        </div>
-      )
-
-      const position = await PositionLocator.find()
-      const style = getComputedStyle(position.getDOMNode())
-
-      expect(style.display).to.equal('inline-block')
-    })
-
-    it('should apply "block"', async () => {
-      await mount(
-        <div style={{ padding: '50px' }}>
-          <div style={{ ...parentDefaults }}>
-            <Position
-              constrain="window"
-              renderTarget={<button>Target</button>}
-              containerDisplay="block"
-            >
-              <div id="content">
-                <div>Content</div>
-              </div>
-            </Position>
-          </div>
-        </div>
-      )
-
-      const position = await PositionLocator.find()
-      const style = getComputedStyle(position.getDOMNode())
-
-      expect(style.display).to.equal('block')
-    })
-  })
-})
